Add v1 API route handling, :8080/v1/api-docs
